refactor(admin): split ModelForm into schema, logic, and UI files
Split the 1258-line ModelForm.tsx into three focused files:

- modelConfigSchemas.ts (166 lines): Provider schema reference data
  and getModelConfigSchemaReference helper.

- modelConfigFormLogic.ts (604 lines): Form state types, empty state
  constant, parse helpers, extractModelConfigFormState, and
  buildModelConfigFromForm. Deduplicates the openrouter/vercel switch
  cases via a shared buildReasoningProviderOptions helper.

- ModelForm.tsx (503 lines): React component, props type, and JSX only.

Also updates ModelConfigFields.tsx to import types from the new
modelConfigFormLogic module.
